ENCD-4137 Move file external creds to helpers dir
