+++ src/encoded/helpers/__init__.py
@@ -1,0 +1,1 @@
+from .upload_credentials import external_creds
